refactor onboarding layout
